requirements.txt
Flask
Pillow
python-dotenv
google-generativeai